requirements.txt
Django==2.0.1
requests==2.18.4
beautifulsoup4==4.6.0
bsdiff4==1.1.4
psycopg2==2.7.3.2
django-debug-toolbar==1.9.1
html5lib==1.0.1
djangorestframework==3.7.7
Keras==2.1.2
# tensorflow-gpu==1.4.1
tensorflow==1.4.1
h5py==2.7.1
Pillow==5.0.0
aiohttp==2.3.8
